+++ CHANGES.md
@@ -1,4 +1,7 @@
-1.4.0 ()
+1.5.0 (unreleased)
+=====
+
+1.4.0 (29-09-2019)
 =====
 
 New:
@@ -8,6 +11,8 @@ New:
 - Added support for deconstructing tuples: `let (z,t,_) = x` (#838)
 - Added `input.{file,harbor}.hls` to read HLS stream (#59, #295, #296).
 - Added `output.hls` to natively stream in HLS (#758).
+- Added `%ffmpeg` native encoder, only for audio encoding for now (#952)
+- Added ffmpeg-based stream decoder, limited to mime type `application/ffmpeg` for now.
 - Added `(to_){string,float,int,bool}_getter` operators to handle getters in
   script side.
 - Made `p` parameter in `smooth_add` a `float` getter (#601)
@@ -37,7 +42,6 @@ New:
 - Added support for bash completion.
 - Added `video.add_text.native`.
 - Added `configure.bindir`
-- Added ffmpeg-based stream decoder, limited to mime type `application/ffmpeg` for now.
 - Added `for` and `while` loop functions.
 - Added `list.case`.
 - Added `metadata.string_getter` and `metadata.float_getter`.
@@ -46,6 +50,8 @@ New:
 - Added `{input,output}.srt` (#898)
 - Added `path.remove_extension`.
 - Added SSL read/write timeout options, use it for incoming socket connections (#932)
+- Added ffmpeg resampler (#947).
+- Added `lsl` and `lsr`.
 
 Changed:
 
@@ -109,6 +115,10 @@ Changed:
   component.
 - Renamed `quote` to `string.quote`.
 - Added `phase_inversion={true/false}` to `%opus` encoder (#937)
+- Fixed encoders forcing frame rate and audio channels too early (#933)
+- Change filename to a string getter in file-based outputs. (#198)
+- Changed `audio.converter.samplerate.preferred` option to
+  `audio.converter.samplerate.converters` to give a list of possible converters.
 
 Fixed:
 
@@ -136,6 +146,7 @@ Fixed:
   reversing the list (#922).
 - File descriptor leak when using openssl-based operators.
 - Fixed SSL read taking too long to timeout (#932)
+- Fixed output starting when underlying source is not available (#393)
 
 1.3.7 (09-04-2019)
 =====